+++ src/extra/locales/ua/translation.json
@@ -18,15 +18,20 @@
         "error": "Нажаль сталася помилка. Вас буде перенаправлено на вікно реєстрації",
         "empty-chat": "Повідомлень ще немає...",
         "no-messages": "Нажаль повідомлень ще немає(",
-        "start-messages" : "Почнемо спілкування!",
-        "join-start" : "Приєднатися і почати спілкування",
-        "incognito" : "Інкогніто",
-        "new-messages" : "Нові повідомлення",
-        "delete" : "Видалити",
-        "edit" : "Редагувати",
-        "editMess" : "ред.  ",
+        "start-messages": "Почнемо спілкування!",
+        "join-start": "Приєднатися і почати спілкування",
+        "incognito": "Інкогніто",
+        "new-messages": "Нові повідомлення",
+        "delete": "Видалити",
+        "edit": "Редагувати",
+        "editMess": "ред.  ",
         "error-empty-mes": "Не можна надсилати пусте повідомлення",
-        "error-long-mes": "Максимум 4000 символів, у Вас "
+        "error-long-mes": "Максимум 4000 символів, у Вас ",
+        "deleted-message": "Видалене повідомлення",
+        "delete-message": "Видалити повідомлення",
+        "want-delete": "Ви бажаєте видалити повідомлення...",
+        "forMe": "Для себе",
+        "forAll": "Для всіх"
     },
     "settingsUser": {
         "settings": "Налаштування профілю",
@@ -55,7 +60,6 @@
         "deleteFolder": "Ви остаточно хочете видалити дану папку?",
         "warningDelete": "Не можна створити більше 10 папок",
         "understand": "Зрозуміло"
-
     },
     "addChat": {
         "error": "Поле обов'язкове для заповнення",
