+++ src/styles/chat-window/messages/mes-another.scss
@@ -79,7 +79,6 @@
                 // .another__reply__img
                 background-color: var(--color-lavender-white);
                 width: 4px;
-                height: 44px;
             }
 
             &__info {
