+++ src/styles/chat-window/messages/mes-self.scss
@@ -75,7 +75,6 @@
                 // .reply__img
                 background-color: #FFF44F;
                 width: 4px;
-                // height: 44px;
             }
 
             &__info {
